# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/transaction/cancel-debt-reminder/:id": {
            "put": {
                "description": "cancel a debt reminder from source or target user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Debt reminder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of debt reminder",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DebtReminderReplyRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/debt-reminder": {
            "post": {
                "description": "Add new Debt reminder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Debt reminder",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DebtReminderRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/external-transfer": {
            "post": {
                "description": "Verify OTP and transaction from internal account to external account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "External transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TransferRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-entity_Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/internal-transfer": {
            "post": {
                "description": "Verify OTP and transaction from internal account to internal account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Internal transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TransferRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-entity_Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/pre-debt-transfer": {
            "post": {
                "description": "Pre Transaction for debt reminder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Debt transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PreDebtTransferRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/pre-external-transfer": {
            "post": {
                "description": "Pre Transaction from internal account to external account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "External transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PreExternalTransferRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/pre-internal-transfer": {
            "post": {
                "description": "Pre Transaction from internal account to internal account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Internal transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PreInternalTransferRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/received-debt-reminder": {
            "get": {
                "description": "get list Receive reminder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Debt reminder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-model_DebtReminderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        },
        "/transaction/sent-debt-reminder": {
            "get": {
                "description": "get list Sent reminder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Debt reminder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization: Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-model_DebtReminderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpcommon.HttpResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bankId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isSourceFee": {
                    "type": "boolean"
                },
                "sourceAccountNumber": {
                    "type": "string"
                },
                "sourceBalance": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "targetAccountNumber": {
                    "type": "string"
                },
                "targetBalance": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "httpcommon.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpcommon.HttpResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpcommon.Error"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpcommon.HttpResponse-entity_Transaction": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Transaction"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpcommon.Error"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpcommon.HttpResponse-model_DebtReminderResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.DebtReminderResponse"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpcommon.Error"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpcommon.HttpResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpcommon.Error"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.DebtReminderReplyRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "model.DebtReminderRequest": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "sourceAccountNumber",
                "targetAccountNumber",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "sourceAccountNumber": {
                    "type": "string"
                },
                "targetAccountNumber": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.DebtReminderResponse": {
            "type": "object",
            "required": [
                "debtReminder",
                "receiver",
                "sender"
            ],
            "properties": {
                "debtReminder": {
                    "$ref": "#/definitions/entity.Transaction"
                },
                "receiver": {
                    "type": "string"
                },
                "reply": {
                    "$ref": "#/definitions/entity.DebtReply"
                },
                "sender": {
                    "type": "string"
                }
            }
        },
...
        "model.PreDebtTransferRequest": {
            "type": "object",
            "required": [
                "transactionId"
            ],
            "properties": {
                "transactionId": {
                    "type": "string"
                }
            }
        },
        "model.PreExternalTransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "isSourceFee",
                "partnerBankId",
                "sourceAccountNumber",
                "targetAccountNumber",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "isSourceFee": {
                    "type": "boolean"
                },
                "partnerBankId": {
                    "type": "integer"
                },
                "sourceAccountNumber": {
                    "type": "string"
                },
                "targetAccountNumber": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.PreInternalTransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "isSourceFee",
                "sourceAccountNumber",
                "targetAccountNumber",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "isSourceFee": {
                    "type": "boolean"
                },
                "sourceAccountNumber": {
                    "type": "string"
                },
                "targetAccountNumber": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.TransferRequest": {
            "type": "object",
            "required": [
                "otp",
                "transactionId"
            ],
            "properties": {
                "otp": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
...
    }
}